requirements.txt
fastapi
uvicorn[standard]
openai>=1.40.0
python-dotenv
pydantic
httpx